eshop_backend/requirements.txt
Django==4.0.6
djangorestframework==3.13.1
django-cors-headers==3.13.0
gunicorn==20.1.0
python-environ==0.4.54
djangorestframework-simplejwt